refactor: consolidate purchasing agent code and remove redundant files
