+++ purchasing_concierge/remote_agent_connection.py
@@ -24,12 +24,10 @@ class RemoteAgentConnections:
     """A class to hold the connections to the remote agents."""
 
     def __init__(self, agent_card: AgentCard, agent_url: str):
-        print(f'agent_card: {agent_card}')
-        print(f'agent_url: {agent_url}')
+        print(f"agent_card: {agent_card}")
+        print(f"agent_url: {agent_url}")
         self._httpx_client = httpx.AsyncClient(timeout=30)
-        self.agent_client = A2AClient(
-            self._httpx_client, agent_card, url=agent_url
-        )
+        self.agent_client = A2AClient(self._httpx_client, agent_card, url=agent_url)
         self.card = agent_card
 
     def get_agent(self) -> AgentCard:
