+++ remote_seller_agents/burger_agent/agent.py
@@ -14,7 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from typing import Literal
 from pydantic import BaseModel
 import uuid
 from crewai import Agent, Crew, LLM, Task, Process
@@ -27,6 +26,7 @@ load_dotenv()
 
 litellm.vertex_project = os.getenv("GCLOUD_PROJECT_ID")
 litellm.vertex_location = os.getenv("GCLOUD_LOCATION")
+
 
 class OrderItem(BaseModel):
     name: str
