+++ test_a2a_client.py
@@ -10,12 +10,12 @@ from a2a.types import (
     AgentCard,
     MessageSendParams,
     SendMessageRequest,
-    SendStreamingMessageRequest,
 )
 
+
 async def main() -> None:
-    PUBLIC_AGENT_CARD_PATH = '/.well-known/agent.json'
-    EXTENDED_AGENT_CARD_PATH = '/agent/authenticatedExtendedCard'
+    PUBLIC_AGENT_CARD_PATH = "/.well-known/agent.json"
+    EXTENDED_AGENT_CARD_PATH = "/agent/authenticatedExtendedCard"
 
     # Configure logging to show INFO level messages
     logging.basicConfig(level=logging.INFO)
@@ -23,7 +23,7 @@ async def main() -> None:
 
     # --8<-- [start:A2ACardResolver]
 
-    base_url = 'http://localhost:10000'
+    base_url = "http://localhost:10000"
 
     async with httpx.AsyncClient(timeout=httpx.Timeout(timeout=30)) as httpx_client:
         # Initialize A2ACardResolver
@@ -39,80 +39,77 @@ async def main() -> None:
 
         try:
             logger.info(
-                f'Attempting to fetch public agent card from: {base_url}{PUBLIC_AGENT_CARD_PATH}'
+                f"Attempting to fetch public agent card from: {base_url}{PUBLIC_AGENT_CARD_PATH}"
             )
             _public_card = (
                 await resolver.get_agent_card()
             )  # Fetches from default pu`b`lic path
-            logger.info('Successfully fetched public agent card:')
-            logger.info(
-                _public_card.model_dump_json(indent=2, exclude_none=True)
-            )
+            logger.info("Successfully fetched public agent card:")
+            logger.info(_public_card.model_dump_json(indent=2, exclude_none=True))
             final_agent_card_to_use = _public_card
             logger.info(
-                '\nUsing PUBLIC agent card for client initialization (default).'
+                "\nUsing PUBLIC agent card for client initialization (default)."
             )
 
             if _public_card.supports_authenticated_extended_card:
                 try:
                     logger.info(
-                        f'\nPublic card supports authenticated extended card. Attempting to fetch from: {base_url}{EXTENDED_AGENT_CARD_PATH}'
+                        f"\nPublic card supports authenticated extended card. Attempting to fetch from: {base_url}{EXTENDED_AGENT_CARD_PATH}"
                     )
                     auth_headers_dict = {
-                        'Authorization': 'Bearer dummy-token-for-extended-card'
+                        "Authorization": "Bearer dummy-token-for-extended-card"
                     }
                     _extended_card = await resolver.get_agent_card(
                         relative_card_path=EXTENDED_AGENT_CARD_PATH,
-                        http_kwargs={'headers': auth_headers_dict},
+                        http_kwargs={"headers": auth_headers_dict},
                     )
                     logger.info(
-                        'Successfully fetched authenticated extended agent card:'
+                        "Successfully fetched authenticated extended agent card:"
                     )
                     logger.info(
-                        _extended_card.model_dump_json(
-                            indent=2, exclude_none=True
-                        )
+                        _extended_card.model_dump_json(indent=2, exclude_none=True)
                     )
                     final_agent_card_to_use = (
                         _extended_card  # Update to use the extended card
                     )
                     logger.info(
-                        '\nUsing AUTHENTICATED EXTENDED agent card for client initialization.'
+                        "\nUsing AUTHENTICATED EXTENDED agent card for client initialization."
                     )
                 except Exception as e_extended:
                     logger.warning(
-                        f'Failed to fetch extended agent card: {e_extended}. Will proceed with public card.',
+                        f"Failed to fetch extended agent card: {e_extended}. Will proceed with public card.",
                         exc_info=True,
                     )
-            elif (
-                _public_card
-            ):  # supports_authenticated_extended_card is False or None
+            elif _public_card:  # supports_authenticated_extended_card is False or None
                 logger.info(
-                    '\nPublic card does not indicate support for an extended card. Using public card.'
+                    "\nPublic card does not indicate support for an extended card. Using public card."
                 )
 
         except Exception as e:
             logger.error(
-                f'Critical error fetching public agent card: {e}', exc_info=True
+                f"Critical error fetching public agent card: {e}", exc_info=True
             )
             raise RuntimeError(
-                'Failed to fetch the public agent card. Cannot continue.'
+                "Failed to fetch the public agent card. Cannot continue."
             ) from e
 
         # --8<-- [start:send_message]
         client = A2AClient(
             httpx_client=httpx_client, agent_card=final_agent_card_to_use
         )
-        logger.info('A2AClient initialized.')
+        logger.info("A2AClient initialized.")
 
         send_message_payload: dict[str, Any] = {
-            'message': {
-                'role': 'user',
-                'parts': [
-                    {'kind': 'text', 'text': 'ah sorry, I want to add 1 bbq chicken too'}
+            "message": {
+                "role": "user",
+                "parts": [
+                    {
+                        "kind": "text",
+                        "text": "ah sorry, I want to add 1 bbq chicken too",
+                    }
                 ],
-                'messageId': uuid4().hex,
-                'contextId': "user123"
+                "messageId": uuid4().hex,
+                "contextId": "user123",
             },
         }
         request = SendMessageRequest(
@@ -120,7 +117,7 @@ async def main() -> None:
         )
 
         response = await client.send_message(request)
-        print(response.model_dump(mode='json', exclude_none=True))
+        print(response.model_dump(mode="json", exclude_none=True))
         # --8<-- [end:send_message]
 
         # --8<-- [start:send_message_streaming]
@@ -136,7 +133,7 @@ async def main() -> None:
         # --8<-- [end:send_message_streaming]
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     import asyncio
 
     asyncio.run(main())
